fix create ingredient api fetch
